+++ src/main/kotlin/com/relogging/server/service/crawling/CrawlingService.kt
@@ -1,4 +1,4 @@
-package com.relogging.server.service
+package com.relogging.server.service.crawling
 
 interface CrawlingService {
     fun crawlAndSaveNewsArticles(): Int
